requirements.txt
joblib==1.5.0
numpy==2.2.5
opencv-python==4.11.0.86
pillow==11.2.1
scikit-learn==1.6.1
scipy==1.15.3
setuptools==78.1.1
threadpoolctl==3.6.0
tqdm==4.67.1
wheel==0.45.1
torch~=2.7.0
torchvision~=0.22.0
pandas~=2.2.3
ultralytics~=8.3.131
pyarrow~=20.0.0